requirements.txt
langchain==0.3.27
faiss-cpu==1.12.0
langchain-community==0.3.31
langchain-experimental==0.3.4
open-clip-torch==3.2.0
pillow==11.3.0
torch==2.8.0
matplotlib==3.10.7
transformers==4.57.0
accelerate==1.10.1
psutil==7.1.0
nltk==3.9.2
bert-score==0.3.13
rouge-score==0.1.2
sacrebleu==2.5.1
datasets==4.2.0




